refactor: move service initialization and routes to Server class
